trocha parametrizace offsetu a force_black, v_co je double

+++ exporter_svg.cpp
@@ -66,7 +66,7 @@ void exporter_svg::write_line(const v_line &line) { // Write one `v_line' in svg
 	}
 	color /= count;
 	if (line.get_type() == v_line_type::stroke) {
-		fprintf(fd, "\"\n       style=\"fill:none;stroke:#%02x%02x%02x;stroke-width:%fpx;stroke-linecap:round;stroke-linejoin:round;stroke-opacity:%f\" />\n", color.val[0], color.val[1], color.val[2], width/count, opacity/count); // write etyle
+		fprintf(fd, "\"\n       style=\"fill:none;stroke:#%02x%02x%02x;stroke-width:%fpx;stroke-linecap:round;stroke-linejoin:round;stroke-opacity:%f\" />\n", (int) color.val[0], (int) color.val[1], (int) color.val[2], width/count, opacity/count); // write etyle
 		if (line.get_group() == v_line_group::group_last) {
 			fprintf(fd, "    </g>\n"); // end group
 		}
@@ -76,7 +76,7 @@ void exporter_svg::write_line(const v_line &line) { // Write one `v_line' in svg
 			group_col = color; // save color of first line
 		}
 		if ((line.get_group() == v_line_group::group_normal) || (line.get_group() == v_line_group::group_last)) {
-			fprintf(fd, " Z\"\n       style=\"fill:#%02x%02x%02x;stroke:none\" />\n", group_col.val[0], group_col.val[1], group_col.val[2]); // filled regions are closed (Z) and have no stroke.
+			fprintf(fd, " Z\"\n       style=\"fill:#%02x%02x%02x;stroke:none\" />\n", (int) group_col.val[0], (int) group_col.val[1], (int) group_col.val[2]); // filled regions are closed (Z) and have no stroke.
 		}
 	}
 }
